aldonis seninterrompan integritan testadon

+++ datumoj.php
@@ -9,12 +9,18 @@ declare(strict_types=1);
 
 chdir(dirname(__FILE__));
 
-file_put_contents('piv2.xml', file_get_contents(
-    'http://kursoj.pagesperso-orange.fr/piv2/piv2.xml'
-));
+if (file_exists('datumoj.db')) {
+    exit;
+}
+
+if (!file_exists('piv2.xml')) {
+    file_put_contents('piv2.xml', file_get_contents(
+        'http://kursoj.pagesperso-orange.fr/piv2/piv2.xml'
+    ));
+}
 $xml = simplexml_load_file('piv2.xml');
 
-$db = new PDO("sqlite:datumoj.db");
+$db = new PDO('sqlite:datumoj.db');
 $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 $db->query('PRAGMA encoding = "UTF-8"');
 $db->query('CREATE TABLE vortoj (vorto TEXT)');
